[FE] rename: 일정 생성 css 파일명 변경
